axum-routes-macros: support for conditional compilation
we have to specifically support conditional compilation attributes
because we need to not simply add them in the enum definition, but also
handle whether or not it should appear when resolving routes and adding
routes in axum::Router.
This is a rough first implementation that unlocks most of use cases.